requirements.txt
mysql-connector-python==9.1.0
pandas==2.1.4
requests==2.28.0